# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Lister tous les messages",
                "operationId": "4216f3eedb7be911361f1b91030253a8",
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Créer un nouveau message",
                "operationId": "c85a9171473a13f151a8672235e01559",
                "requestBody": {
                    "description": "Données du message",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message créé avec succès"
                    }
                }
            }
        },
        "/api/messages/{id}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Récupérer un message par ID",
                "operationId": "2d72cf14bf2d55c9c9dfd297da04d404",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du message",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "404": {
                        "description": "Message non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Messages"
                ],
                "summary": "Mettre à jour un message",
                "operationId": "f5c039674f51b998b915d3172b879661",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du message",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Données du message à mettre à jour",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message mis à jour avec succès"
                    },
                    "404": {
                        "description": "Message non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Messages"
                ],
                "summary": "Supprimer un message",
                "operationId": "3900a1f5d7b34a2d1a02b42535a3fdd9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du message",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Message supprimé avec succès"
                    },
                    "404": {
                        "description": "Message non trouvé"
                    }
                }
            }
        },
        "/api/profiles": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Lister tous les profils",
                "operationId": "3209d3b3ace427ff126d7e173f0e6759",
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Créer un nouveau profil",
                "operationId": "b8873d217549ada3677c5334401e5622",
                "requestBody": {
                    "description": "Données du profil",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Profile"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Profil créé avec succès"
                    }
                }
            }
        },
        "/api/profiles/{id}": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Récupérer un profil par ID",
                "operationId": "4c97aadaa22fadcd119f2cc6e55cf624",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du profil",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profil ou utilisateur non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Mettre à jour un profil",
                "operationId": "16d08e84b8ee4ad6e1fa16bef8a7baac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du profil",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Données du profil à mettre à jour",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Profile"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil mis à jour avec succès"
                    },
                    "404": {
                        "description": "Profil non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Supprimer un profil",
                "operationId": "3b8a8572606da7a63904b9f1d1c0d4cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du profil",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Profil supprimé avec succès"
                    },
                    "404": {
                        "description": "Profil non trouvé"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Message": {
                "properties": {
                    "match_id": {
                        "type": "integer",
                        "example": ""
                    },
                    "sender_id": {
                        "type": "integer",
                        "example": ""
                    },
                    "content": {
                        "type": "string",
                        "example": ""
                    },
                    "sent_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": ""
                    }
                },
                "type": "object"
            },
            "Profile": {
                "properties": {
                    "firstname": {
                        "type": "string",
                        "example": ""
                    },
                    "lastname": {
                        "type": "string",
                        "example": ""
                    },
                    "location": {
                        "type": "string",
                        "example": ""
                    },
                    "interests": {
                        "type": "string",
                        "example": ""
                    },
                    "bio": {
                        "type": "string",
                        "example": ""
                    },
                    "profile_picture": {
                        "type": "string",
                        "example": ""
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}